backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.10
celery==5.5.2
django-celery-beat==2.5.0
django-cors-headers==4.3.1